fix: pop up not shown when progress bar is resized
When this.gantt.bar_being_dragged is compared to null, there are no problems with showing pop up after resizing progress bar.
	if (this.gantt.bar_being_dragged == null) return;

In set_action_completed(), it seems that the delay in setTimeout() is too big. By reducing it, the user can see the pop up after resizing bars with no bugs.
	Values changed -> From 1000 to 200