Use CSS instead of JS to align lower text dates
This drastically improves performance of the `create_dates` method, which is important when rendering a many dates (i.e. for the `Hours` view).

The previous implementation calls `clientWidth` in a loop which was over 50% of the execution time on this method.

In my tests, while rendering a chart with tasks spanning approximately 1 year, the execution time of `create_dates` decreased from 27,844ms to 6,342ms when switching to the hours view